chg: proper util for state sync
